folder structure added for dozen email automation system
