+++ src/configurations/config.py
@@ -1,0 +1,4 @@
+# # Public GoogleSheets URL - not secure!
+# SHEET_ID = "18iZ2cDnvyaNpjytD-NHZykfSTQECMiidV-cW8c2vAVQ"  # !!! CHANGE ME !!!
+# SHEET_NAME = "Invoice-client-list"  # !!! CHANGE ME !!!
+URL = f"https://docs.google.com/spreadsheets/d/18iZ2cDnvyaNpjytD-NHZykfSTQECMiidV-cW8c2vAVQ/gviz/tq?tqx=out:csv&sheet=Invoice-client-list"
