+++ src/operators/email_sender.py
@@ -1,0 +1,19 @@
+from datetime import date  
+from src.connector.send_email import send_email 
+
+def query_data_and_send_emails(df):
+    """Query the DataFrame and send email reminders."""
+    present = date.today()
+    email_counter = 0
+    for _, row in df.iterrows():
+        if (present >= row["reminder_date"].date()) and (row["has_paid"] == "no"):
+            send_email(
+                subject=f'[Invoice Reminder] Invoice: {row["invoice_no"]}',
+                receiver_email=row["email"],
+                name=row["name"],
+                due_date=row["due_date"].strftime("%d, %b %Y"),  # e.g., 11, Aug 2022
+                invoice_no=row["invoice_no"],
+                amount=row["amount"],
+            )
+            email_counter += 1
+    return f"Total Emails Sent: {email_counter}"
